requirements.txt
Flask==2.1.0
python-telegram-bot
google-cloud-vision
requests
pillow